Refactor and code cleanup.

+++ ecommerce.api/Managers/ProductDataManager.cs
@@ -1,7 +1,6 @@
 using ecommerce.api.Classes;
 using ecommerce.api.Data;
 using ecommerce.api.Entities;
-using ecommerce.api.Services.Mappers;
 using Microsoft.EntityFrameworkCore;
 
 namespace ecommerce.api.Managers;
@@ -49,7 +48,7 @@ public class ProductDataManager
         return categories;
     }
 
-    public async Task<ProductEntity> GetProduct(Guid id)
+    public async Task<ProductEntity?> GetProduct(Guid id)
     {
         var product = await _context.Products
             .Include(p => p.Images)
@@ -63,12 +62,12 @@ public class ProductDataManager
     {
         var mappedProduct = new ProductEntity()
         {
-            Id = product.Id,
             Name = product.Name,
             Description = product.Description,
             Category = product.Category,
+            Color = product.Color,
             PricePerUnit = product.PricePerUnit,
-            Discount = product.Discount
+            Discount = product.Discount,
         };
         
         await _context.Products.AddAsync(mappedProduct);
@@ -77,28 +76,36 @@ public class ProductDataManager
         return mappedProduct;
     }
 
-    public async Task<ProductEntity> UpdateProduct(ProductModel product)
+    public async Task<ProductEntity?> UpdateProduct(ProductModel product)
     {
         var existingProduct = await _context.Products
             .Include(p => p.Images)
             .FirstOrDefaultAsync(p => p.Id == product.Id);
+
+        if (existingProduct == null)
+            return null;
         
         existingProduct.Name = product.Name;
         existingProduct.Description = product.Description;
         existingProduct.Category = product.Category;
+        existingProduct.Color = product.Color;
         existingProduct.PricePerUnit = product.PricePerUnit;
         existingProduct.Discount = product.Discount;
+        
+        existingProduct.UpdatedDate = DateTime.UtcNow;
         
         await _context.SaveChangesAsync();
 
         return existingProduct;
     }
 
-    public async void DeleteProduct(Guid id)
+    public async Task DeleteProduct(Guid id)
     {
         var product = await _context.Products.FirstOrDefaultAsync(p => p.Id == id);
 
+        if (product == null) return;
+        
         _context.Products.Remove(product);
-        _context.SaveChanges();
+        await _context.SaveChangesAsync();
     }
 }
